Allow token authentication to websocket, Mk. 2
This is attempt number 2 at #3345, which I screwed up first time round
by accidentally disabling session authentication to the WebSocket.

This uses `pyramid-multiauth` from Mozilla to "stack" the newly-added
TokenAuthenticationPolicy and Pyramid's existing
SessionAuthenticationPolicy. In the words of the `pyramid-multiauth`
documentation, stacking auth policies means that "the results from the
stacked authentication policies are combined as follows":

- authenticated_userid: return userid from first successful policy
- unauthenticated_userid: return userid from first successful policy
- effective_principals: return union of principals from all policies
- remember: return headers from all policies
- forget: return headers from all policies

This means that it is possible to authenticate to the websocket with
either (or both) an API token and a valid session cookie, preserving the
behaviour the client depends on, while making 3rd-party integration with
the websocket easier.