List pull requests that were merged without a merge commit in list-changes-since.py
The changelog helper script relied on parsing merge commit messages in order to
get the numbers of pull requests merged since a given tag.

This did not include pull requests that were merged via a fast-forward merge.

Fix this by using the GitHub API to enumerate pull requests in descending
last-update order and finding those merged since the tag was created.

There is a caveat that this relies on the system time being reasonably accurate
on the system where the tag was created.