Clean up admin "admins" view and associated tests
This commit is mainly about cleaning up the tests for the views
responsible for managing which users are admins so that they are
consistent with the test changes made in #3424.

It also makes the `admins_add` and `admins_remove` views consistent in
their behaviour (they now both do redirect-after-post rather than just
one of them), and fixes a bug that allowed the `admins_remove` view to
crash if the username provided to it was invalid.