Remove the h.accounts.make_staff helper
As part of the work started in #3423, I'm working up to making
the `User.get_by_username` classmethod take an explicit database session
rather than using the query property.

There are a couple of places that use `User.get_by_username` that don't
currently have access to the request object, including these helpers
(`make_admin` and `make_staff`) in `h.accounts`.

So, my options are:

1. Pass the request object into these helpers.
2. Remove the layer of indirection that the helpers represent.

I've opted here for the latter, because they don't really seem to be
doing anything particularly useful. There is only one place that
`make_staff` is used, namely the admin panel. The code needed to handle
the NoSuchUserError is almost exactly as verbose as the code needed
without this helper.

Moreover, in the admin views, the helper can only be used to make a user
staff, not to remove their staff status, so the companion view
(`staff_remove`) already knows about the `User` model.

This commit entails a major cleanup of the relevant view tests in line
with the changes made in #3424 and #3425.