Do not use query property in Blocklist model
This commit is part of an effort to standardise all ORM classmethods so
they take a database session as their first parameter and do not rely on
a thread-local query property.

The Blocklist model stores a list of patterns to be matched against
incoming URLs for badge requests to see if the badge count should be
forced to 0 for that request.

There were a number of classmethods on the Blocklist model, most of
which were there to assist the admin views for updating the blocklist,
rather than for the primary badge endpoint. Most of these are now gone.

Specifically:

- The only classmethod remaining is Blocklist.is_blocked, which now
  takes a database session as its first parameter.
- Blocklist.all was an alias for Blocklist.query.all... now gone.
- Blocklist.get_by_uri was only used by a) a validator that wasn't
  concurrency-safe, and b) Blocklist.is_blocked. I've removed the
  validator and replaced the relevant validation with a concurrency-safe
  version, and moved the lookup query directly into .is_blocked.

While in this part of the code, I've made the admin POST view functions
do a redirect-after-post to prevent double submission problems.

Lastly, I've rejigged the tests for the admin views in much the same way
as I did in #3424 and #3425 for the admins/staff admin views.