Initialize `tags` and `text` fields of all annotations loaded into the app
Several places in the <annotation> component assumed that the `tags`
field was an array, as did <tag-editor>. The `tags` and `text` fields
are always set by the server for annotations received from the API but
are not set by Annotator when creating new annotations in the page.

Initializing these fields in the reducer, which is the only place that
the loaded set of annotations can be modified, means that the rest of
the app can rely on them having the expected type.

This fixes an error when saving new annotations.