Make the standalone annotation page work for replies
When displaying a standalone annotation page for replies, display the
entire conversation thread which that reply is in, including the
original annotation.

Previously standalone annotation pages for replies were broken because
only the reply itself was fetched, the the existence of the top-level
annotation was inferred from the reply's `references` field but because
it was missing, a blank "Message not available" card was shown instead.

Showing the entire thread fixes this problem and also allows the user to
see the complete context of the conversation.

Also refactor the tests to be less tied to the implementation details of
the class and expressed more in terms of what the view does from a
user's point of view.

Fixes #3367