Rewrite factories.Annotation to use Postgres
Rewrite factories.Annotation to generate the new h.api.models.Annotation
objects that use the new PostgreSQL annotation storage, instead of
generating legacy Elasticsearch storage objects.

In the process also add Document, DocumentURI and DocumentMeta factories
that the Annotation factory uses, these can also be used independently.

Also replace factories_test.py with a much simpler set of tests that are
really just intended to break if the factories are completely broken,
hopefully helping to make it clear that it's the factories that are
broken (and not all the other tests that would be failing).

There's not much too the various factories that isn't just declarative
and getting factory_boy to do things for us, so not much point in us
testing the behaviour in detail.

Some parts such as Annotation.make_metadata() probably cross this line
but would probably require some pretty awkward mocking to test because
they involve so much randomness.