Add migration script to fix document_uri unique constraint
This commit adds a migration script that adds NOT NULL constraints and
default values to the type and content_type columns of the document_uri
table, thus making the unique constraint on that table works as
intended.

It also deduplicates any existing data in the table to make it
compatible with the fixed unique constraint.

Note that to enable the tests for the migration script that this commit
includes, this commit does not update the database schema in the
models.py file to match that which the migration script creates yet.

The models.py change, and also removing of the migration script tests
which will no longer work after the models.py change, will be done in
the following commit.