Fix ConcurrentUpdateError when creating new annotations
If parse_document_claims returns 'type': None and/or 'content_type': None in
any of its document uri dicts then when document.py later uses those
dicts to search the db for pre-existing DocumentURI's it will search for
DocumentURI's with type (or content_type) None.

In fact there's a default value of '' on type and content_type in the
DocumentURI ORM class, so if you did DocumentURI(type=None) then by the
time it gets committed to the db type will have become ''.

So document.py's search will not find any existing DocumentURI's and
it'll go ahead and create a new DocumentURI(type=None) and add it to the
db.

On flush or commit this type=None will become '' and clash with the
pre-existing DocumentURI with type='', violating the unique constraint
on the document_uri db table and crashing with IntegrityError.

This gets turned into a ConcurrentUpdateError causing pyramid_tm to
replay the request but the problem in fact has nothing to do with
concurrency - the same exact thing will happen again and
ConcurrentUpdateError will be raised again.

After three ConcurrentUpdateErrors pyramid_tm gives up and the request
crashes.