Add a service to manage group creation and membership
In preparation for some substantial reorganisation of the groups views
(which in turn is in preparation for changes which we'll be making to
groups in the near future) this commit adds a GroupsService which hides
some of the details of creating groups and managing their membership.

In this commit the service remains unused.