Rearrange groups views to take advantage of GroupsService
Now that the business logic associated with administering groups is
mostly encapsulated by the GroupsService, we can substantially simplify
the views code. This is achieved by doing the following:

- Retrieval of the group for join/leave/read views moves into a Pyramid
  traversal operation, so that each view simply receives the group
  object directly, rather than having to fetch it itself.

- Rather than each view needing to do its own
  `if request.authenticated_userid is None:` check, we use the Pyramid
  `effective_principals` predicate in the view config. This also allows
  us to remove a conditional from the read view, by matching different
  view functions dependent on the user's authentication status.

- Relieve the views of the need to do an explicit membership check by
  adding an ACL to the `Group` model. Only members of the group have the
  "read" permission, so we can then use
  `request.has_permission("read")`.

This results in somewhat less mock-heavy test code for views.py, as well
as improved test coverage. Hopefully it will also make it easier to add
and modify functionality in future, too.