Improve error for invalid characters in username
Improve the error message for invalid characters in the username.

The default error message (from colander) is "String does not match
expected pattern" which doesn't explain what the expected pattern is,
and contains the jargon words string and pattern.

Replace this with a jargon-free error message that tells the user which
characters they need to remove.

The new error message follows the style of existing error messages on
the same field, e.g. "Shorter than minimum length 3".