Allow multiple 'uri' parameters to the search query
In order to have a better chance of retrieving all the annotations we
know about for a document, it should be possible to query by all known
URIs for that document.

This commit makes it possible to supply multiple 'uri' parameters to the
API search endpoint. The resulting query will match annotations in which
`target.scope` (i.e. the normalised target URI) matches *any* of the
URIs in the set expanded from *all* of the query parameter values.

This is an entirely backwards compatible change. In the case where only
one `uri` parameter is provided, the query output is identical.