Use request-scoped rather than thread-local session
SQLAlchemy's scoped_session is very nice, but is unnecessary magic in
Pyramid, where it's possible to scope a database session to a request
without too much effort. In addition, scoped_session makes it very hard
to alter the connection on a request-by-request basis (in order, for
example, to guarantee that specific views are read-only) while
preventing inappropriate access to the database session from outside the
request scope -- not something we do currently, but an option I'd like
to keep open for the future.

This commit replaces `h.db.Session` with a stateless session factory,
which can be used to construct sessions but which has no implicit
engine/connection binding.

The engine/connection pool is stored on the application registry, which
enables us to keep the MetaData objects completely stateless.

This commit also adds a "manager hook" to pyramid_tm which ensures
complete isolation of the transaction manager between requests, by
creating a new TransactionManager for each request rather than using the
threadlocal `transaction.manager`.

Finally, this commit also updates the `db_session` fixture in the tests
to use a fast test rollback technique that doesn't require
monkeypatching the session object. See the `db_session` docstring for
more details.