Fix DocumentURI unique constraint migration
This fixes two problems with 467ea2898660_fix_document_uri_unique_constraint.py:

1. When updating a document_uri and changing either its type,
content_type or both from NULL to '' there may already exist an
equivalent row that already has '' instead of NULL. The UPDATE would
crash with IntegrityError. In this case DELETE the row with the NULL(s)
instead of updating it.

2. When doing a DELETE or UPDATE on a table the rows being deleted or
updated are locked for the duration of the transaction. On our
production data this migration ends up deleting or updating most of the
rows in document_uri so most of the table is locked for several seconds.

So batch these DELETEs and UPDATEs into transactions of 25 rows at a
time. This makes the migration script take much longer to run but
ensures that the rows are never locked for too long at once (on my
laptop, on production data, each call to commit() takes less than
0.05s).