Unify the way update and creation of annotations is handled
As the next step towards making Annotation instances immutable, change
creation of annotations on the server in <annotation> to make a copy of
the annotation and call `$create()` on that, in the same way as
`$update()` is handled.

With this change, the app can now rely on reference equality to test
whether an annotation has changed after the annotation card has been
created for the first time.