Use direct links in reply notification emails
In the notification emails that we send when someone replies to
someone's annotation, instead of linking to the individual page of the
reply use a bouncer direct link to the top-level parent annotation of
the reply. (We can't yet do a direct link to the reply itself - the
client doesn't support this.)

The human_timestamp Jinja2 filter is removed because it's no longer used.