Split assets environment in two
The web application adds asset links to pages with the help of an assets
environment (an instance of `h.assets.Environment`). The environment
knows how to map bundle names such as "site_js" to a series of URLs for
inclusion -- it does this by reading two files:

- `assets.ini`, which maps bundle names to a list of individual asset
  filenames.
- `manifest.json`, which maps individual asset filenames to URL paths
  (for the purposes of busting browser caches).

Up until now both client and site assets have been built in the same
step, and have written out a single `manifest.json`. Now, with the
client installed with npm, there are two manifest files: one for the
client and one for the site assets.

This commit splits the rest of the asset handling into two to match this
division. Client asset bundles are now defined in h/assets_client.ini,
and the client assets have a separate environment instance which reads
the manifest from the installed `hypothesis` package in `node_modules/`.

We can remove this additional assets environment and configuration once
two things have happened:

1. Responsibility for generating `app.html` and `embed.js` have moved
   into the client build.
2. The remaining site pages which depend on the client CSS (AKA
   `app_css`) have been moved to site CSS.