Remove digest triggering in AnnotationUISync
This is no longer needed following the addition of middleware to the
Redux store which triggers a digest following a UI state change if
necessary.

Also add a comment in the WebSocket code that we will be able to remove
the $apply() call once session state has been moved to the Redux store.