Use streamer session in realtime message handlers
This commit ensures that all database access inside the realtime message
handlers in `h/streamer/messages.py` is done using the streamer-managed
session created in `process_work_queue`.

This ensures that:

- the underlying connection is returned to the pool after handling each
  message...
- ...and thus we don't accidentally overflow the database connection
  pool when handling lots of messages (or a few messages with lots of
  clients)
- also, the transaction isolation level set in `streamer.py` is
  respected

In order to do this, the created session is now passed into the message
handlers, and from there passed to the subsidiary
`handle_annotation_message` and `handle_user_message` handlers.

Further, and in order to ensure that this problem doesn't recur, I've
removed the `request` property from the `h.websocket.WebSocket`
instances that represent client connections to the WebSocket. This means
that we can't accidentally create a new database connection by accessing
`socket.request.db`. Now the socket instance is informed of its auth
status through the WSGI environment. This has the nice side effect of
removing our last use of `pyramid.threadlocal` from the codebase