Add a links service, to allow generating links to annotations
Generating links to annotations is currently a responsibility of the
`h.api.presenters` module, which uses a passed-in Pyramid request object
to do much of the heavy lifting.

This commit adds a links service (currently unused) which will replace
this functionality which will allow us to remove the need to pass a
request object into the presenters.

This, in turn, will allow us to entirely remove the use of a request
object from the streamer code -- thus eliminating the possibility that
we accidentally use `request.db` (as we currently do), thus opening
unintended connections to the database server.