Remove request dependency from presenter classes
This commit removes the need to pass a Pyramid request object into a
presenter class, which will in turn make it possible to remove all need
for a request object from the streamer message handling code.

Instead, the AnnotationJSONPresenter and AnnotationJSONLDPresenter
classes accept a "links service" argument which they use to generate
links.