Remove referer from groups page links
Ask browsers not to include path part of the URL in the HTTP "Referer"
header in requests sent when clicking on links on the groups page, by
adding a <meta name="referrer" content="origin"> HTML tag to the <head>
of the groups page.

The problem is that some of the links on a group's page - the links to
the documents that the group has annotated - are links to third-party
websites and the values of the Referer header sent to these websites
would be the URLs of the group pages.

If you have the URL of a group's page then you can join the group and
see the group's annotations, and those URLs were being revealed to
anyone with access to the HTTP request logs of the servers of the
annotated documents.

The <meta name="referrer" tag actually asks browsers not to send the
path for _any_ of the links on the page, including links to other pages
of our own hypothes.is website, even though it's only the links to
third-party sites that we don't want to send the path to.

Other possible fixes that were rejected:

- A `Referrer-Policy: origin` HTTP header does the same thing as the
  <meta name="referrer" content="origin"> tag and would be a nicer way
  to do it, but none of Chrome, Firefox, Safari or Edge support the
  Referrer-Policy header (I tested it in Chrome and it didn't seem to
  work, the FastMail blog post linked below states that it doesn't work
  in the other browsers either).

- It's possible to add a rel="noreferrer" attribute to individual links,
  instead of a <meta name="referrer" tag for the whole page, but
  rel="noreferrer" is not as widely supported by browsers as <meta
  name="referrer" is (in particular, rel="noreferrer" is not supported
  by older versions of Firefox or by any version of Internet Explorer).

- It's also possible to use a
  <meta name="referrer" content="origin-when-cross-origin"> tag to tell
  the browser to only send the path in the Referer when following links
  to URLs of the same origin as the current page. But
  "origin-when-cross-origin" is not as widely supported by browsers as
  "origin" (in particular, Edge and Safari don't support it).

Since we're not actually doing anything with the path in the Referer
header when people follow links from the groups page to other pages of
our own site, I judged it better to go for the widest browser support
than to try to only strip the path from the links that we need to strip
it from.

For further reading about stripping Referer see:

https://blog.fastmail.com/2016/06/20/everything-you-could-ever-want-to-know-and-more-about-controlling-the-referer-header/

Other approaches that were considered:

* A link from a group's page to https://example.com/annotated_document
  could link to https://hypothes.is/redirect/https://example.com/annotated_document,
  this /redirect/* page would then simply issue an HTTP redirect to the
  given *. Unfortunately this doesn't work: browsers preserve the
  original Referer when HTTP-redirected.

* You can do tricks with JavaScript redirects, falling back to HTML (not
  HTTP) redirects when JavaScript isn't available, but browser behaviour
  with the various available techniques varies between browsers and it
  seems complicated and hacky.

* The links to documents on the group's page could become bouncer direct
  links. Bouncer seems to strip Referer headers (at least in Chrome).
  But we don't support direct linking to a group's annotations of a
  document (as opposed to an individual annotation yet). Also stripping
  Referer is arguably a bug / unwanted behaviour in bouncer.

* Use separate pages with different URLs for:

  1. The invite link that you need in order to join a group, and
  2. The page that you visit to see what a group has been doing when
     you're already a member of the group

  1 would be the secret link needed to join the group, and would not contain
  any links to third-party sites.

  2 would contain the links to the annotated documents, but would not
  contain the secret code in the URL.

  This seems like it might be a good approach, but is a lot more work
  than just inserting a <meta> tag.

* Have a proper group invite system that doesn't rely on a secret URL.
  Again, too much work for now.