Explicitly escape user text in Deform forms
User text was already being escaped in our Deform forms, but only
because the app's main Jinja2 environment, which our Deform Jinja2
renderer creates its own Jinja2 environment as an overlay of, had
Jinja2's autoescaping enabled.

This commit ensures that our Deform Jinja2 renderer escapes user text
even if the base Jinja2 environment doesn't.

Since the Deform renderer is for rendering forms, it's good for it to be
explicit about how it does escaping, rather than implicitly relying on
the base env to do it.