Add a new "logged-out navbar" component
This can be used simply by putting:

{% include  "h:templates/includes/logged-out-navbar.html.jinja2" %}

in any Jinja2 template.

The CSS for the component is currently in site.scss which means it's
included on every page, whether the page uses this component or not.

TODO: Make the component include its own CSS so that the CSS is only
used on pages where the component is used.