Make site_v2_css the default style bundle
This makes site_v2_css the default style bundle applied by the
`base.html.jinja2` template, and adds opt-outs for those pages
(basically just group pages) which don't already explicitly opt-out of
the default bundle.

This might seem a counterintuitive approach, as this commit means that
the layout "default bundle" is now not being used by *any* of the
templates that extend the layout. But by making the default bundle
`site_v2_css` we make it much easier to grep for templates which need
updating to use the default bundle, and once fixed we simply *remove
overrides* from those templates.